+++ tests/TheTVDbAPITest.php
@@ -16,6 +16,7 @@ declare(strict_types = 1);
 
 namespace CanIHaveSomeCoffee\TheTVDbAPI\Tests;
 
+use CanIHaveSomeCoffee\TheTVDbAPI\TheTVDbAPI;
 use Exception;
 use GuzzleHttp\Client;
 use GuzzleHttp\Exception\RequestException;
@@ -23,7 +24,6 @@ use GuzzleHttp\Handler\MockHandler;
 use GuzzleHttp\HandlerStack;
 use GuzzleHttp\Psr7\Request;
 use GuzzleHttp\Psr7\Response;
-use InvalidArgumentException;
 use CanIHaveSomeCoffee\TheTVDbAPI\Exception\ParseException;
 use CanIHaveSomeCoffee\TheTVDbAPI\Exception\ResourceNotFoundException;
 use CanIHaveSomeCoffee\TheTVDbAPI\Exception\UnauthorizedException;
@@ -33,8 +33,6 @@ use CanIHaveSomeCoffee\TheTVDbAPI\Route\LanguagesRoute;
 use CanIHaveSomeCoffee\TheTVDbAPI\Route\SearchRoute;
 use CanIHaveSomeCoffee\TheTVDbAPI\Route\SeriesRoute;
 use CanIHaveSomeCoffee\TheTVDbAPI\Route\UpdatesRoute;
-use CanIHaveSomeCoffee\TheTVDbAPI\Route\UsersRoute;
-use CanIHaveSomeCoffee\TheTVDbAPI\TheTVDbAPI;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -103,7 +101,7 @@ class TheTVDbAPITest extends TestCase
 
         $handler = HandlerStack::create($mock);
         $client = new Client(['handler' => $handler]);
-        $test_instance = new TheTVDbAPI($client);
+        $test_instance = new TheTVDbAPI("eng", "", $client);
         $test_instance->setToken('ABC');
 
         $headers = $test_instance->requestHeaders('GET', '/');
@@ -113,7 +111,7 @@ class TheTVDbAPITest extends TestCase
     public function testRequestWithMissingToken()
     {
         $client = $this->createTokenErrorClient();
-        $test_instance = new TheTVDbAPI($client);
+        $test_instance = new TheTVDbAPI("eng", "", $client);
         $test_instance->setToken('ABC');
 
         static::expectException(UnauthorizedException::class);
@@ -123,7 +121,7 @@ class TheTVDbAPITest extends TestCase
     public function testRequestHeadersPathError()
     {
         $client = $this->createPathErrorClient();
-        $test_instance = new TheTVDbAPI($client);
+        $test_instance = new TheTVDbAPI("eng", "", $client);
         $test_instance->setToken('ABC');
 
         // Not found
@@ -138,7 +136,7 @@ class TheTVDbAPITest extends TestCase
             new Response(200, [], $input),
             new RequestException("Error Communicating with Server", new Request('GET', 'test'))
         ]);
-        $test_instance = new TheTVDbAPI($client);
+        $test_instance = new TheTVDbAPI("eng", "", $client);
         $test_instance->setToken('ABC');
 
         $response = $test_instance->performAPICall("GET", "/");
@@ -148,7 +146,7 @@ class TheTVDbAPITest extends TestCase
     public function testAPICallWithMissingToken()
     {
         $client = $this->createTokenErrorClient();
-        $test_instance = new TheTVDbAPI($client);
+        $test_instance = new TheTVDbAPI("eng", "", $client);
         $test_instance->setToken('ABC');
 
         static::expectException(UnauthorizedException::class);
@@ -158,7 +156,7 @@ class TheTVDbAPITest extends TestCase
     public function testAPICallPathError()
     {
         $client = $this->createPathErrorClient();
-        $test_instance = new TheTVDbAPI($client);
+        $test_instance = new TheTVDbAPI("eng", "", $client);
         $test_instance->setToken('ABC');
 
         // Not found
@@ -172,7 +170,7 @@ class TheTVDbAPITest extends TestCase
             new Response(201, []),
             new RequestException("Error Communicating with Server", new Request('GET', 'test'))
         ]);
-        $test_instance = new TheTVDbAPI($client);
+        $test_instance = new TheTVDbAPI("eng", "", $client);
         $test_instance->setToken('ABC');
 
         static::expectException(Exception::class);
@@ -186,7 +184,7 @@ class TheTVDbAPITest extends TestCase
             new Response(200, [], $input),
             new RequestException("Error Communicating with Server", new Request('GET', 'test'))
         ]);
-        $test_instance = new TheTVDbAPI($client);
+        $test_instance = new TheTVDbAPI("eng", "", $client);
         $test_instance->setToken('ABC');
 
         static::expectException(ParseException::class);
@@ -200,7 +198,7 @@ class TheTVDbAPITest extends TestCase
             new Response(200, [], $input),
             new RequestException("Error Communicating with Server", new Request('GET', 'test'))
         ]);
-        $test_instance = new TheTVDbAPI($client);
+        $test_instance = new TheTVDbAPI("eng", "", $client);
         $test_instance->setToken('ABC');
 
         static::expectException(ParseException::class);
@@ -221,7 +219,7 @@ class TheTVDbAPITest extends TestCase
             new Response(200, [], json_encode($expected)),
             new RequestException("Error Communicating with Server", new Request('GET', 'test'))
         ]);
-        $test_instance = new TheTVDbAPI($client);
+        $test_instance = new TheTVDbAPI("eng", "", $client);
         $test_instance->setToken('ABC');
 
         $result = $test_instance->performAPICallWithJsonResponse("GET", "/");
@@ -236,7 +234,7 @@ class TheTVDbAPITest extends TestCase
             new Response(200, [], json_encode($expected)),
             new RequestException("Error Communicating with Server", new Request('GET', 'test'))
         ]);
-        $test_instance = new TheTVDbAPI($client);
+        $test_instance = new TheTVDbAPI("eng", "", $client);
         $test_instance->setToken('ABC');
 
         $result = $test_instance->performAPICallWithJsonResponse("GET", "/");
@@ -250,7 +248,7 @@ class TheTVDbAPITest extends TestCase
             new Response(200, [], json_encode($expected)),
             new RequestException("Error Communicating with Server", new Request('GET', 'test'))
         ]);
-        $test_instance = new TheTVDbAPI($client);
+        $test_instance = new TheTVDbAPI("eng", "", $client);
         $test_instance->setToken('ABC');
 
         $result = $test_instance->performAPICallWithJsonResponse("GET", "/");